+++ src/System.Device.Gpio/System/Device/I2c/Devices/Windows10I2cDevice.Linux.cs
@@ -1,0 +1,24 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System.Device.I2c.Drivers
+{
+    public class Windows10I2cDevice : I2cDevice
+    {
+        public Windows10I2cDevice(I2cConnectionSettings settings) =>
+            throw new PlatformNotSupportedException($"The {GetType().Name} class is not available on Linux.");
+
+        public override I2cConnectionSettings ConnectionSettings => throw new PlatformNotSupportedException();
+
+        public override byte ReadByte() => throw new PlatformNotSupportedException();
+
+        public override void Read(Span<byte> buffer) => throw new PlatformNotSupportedException();
+
+        public override void WriteByte(byte value) => throw new PlatformNotSupportedException();
+
+        public override void Write(ReadOnlySpan<byte> buffer) => throw new PlatformNotSupportedException();
+
+        public override void WriteRead(ReadOnlySpan<byte> writeBuffer, Span<byte> readBuffer) => throw new PlatformNotSupportedException();
+    }
+}
